Revert "x2: overlay: use default light sensor debouncing values"
This reverts commit add145272f8b6558ae6df61aa755454549d57e5c.

As mentioned in the original commit we were switching to default debouncing values to fix the "contrast bug" where the contrast would switch from high to normal when going from an app with a bright background to one with a dark background.

It turns out this behaviour is more likely due to either LiveDisplay's "reduce power consumption" mode or something else causing the "contrast bug."

Recommend reverting back to faster debouncing values as the default ones are really slow to react and the faster values don't have any negative effects that I have observed.

Change-Id: I707c3cdaf384e1c6f5c411b7dc107ab6929c4337