Update patch set 2
Patch Set 2:

(1 comment)

Focus issues on x2 is caused by some of the earlier camera modules having really cheap actuators that either fail over time, or if the device is dropped it will fail to achieve focus on the value from 19s kernel source code... even so, I've seen some x2 users with the focus fix value unable to focus unless they do a chopping motion with the phone in their hand which seems to jar the actuator and gets it to function properly.

The issue with trying to force the 16s kernel source camera actuator value on everyone is it seems to jar the actuator more and has a more audible actuator initialization sound (the only way we can tell it works on our devices that don't have focus issues for if the focus fix value is working).

Unfortunately, I don't see a way we can even automatically detect it... hence the toggle value.

Patch-set: 2
CC: Marc Bourgoin <19252@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>