Update patch set 2
Patch Set 2:

For imx230 camera module manufacture date? Mine seems to be the revised date... so anyone with a module date of 2016-07-08 or newer is fine, but earlier and they seem to be the unfortunate owners of an x2 which puts them at risk for focus issues.

Think of it in the sense of the HTC One M7 with the purple tink camera flaw... albeit, that was a more wide-spread hardware defect compared to x2's imx230 camera actuator focus issues...

Patch-set: 2